Finished the first iteration

+++ app/src/main/java/com/eventlocator/eventlocatororganizers/adapters/CanceledEventAdapter.kt
@@ -22,7 +22,7 @@ class CanceledEventAdapter(private val canceledEvents: ArrayList<Event>):
         init{
             binding.root.setOnClickListener {
                 val intent = Intent(context, ViewEventActivity::class.java)
-                intent.putExtra("eventID", binding.tvEventID.text.toString().toInt())
+                intent.putExtra("eventID", binding.tvEventID.text.toString().toLong())
                 context.startActivity(intent)
             }
         }
